requirements.txt
requests
beautifulsoup4
google-genai
streamlit